Update policy file with the correct renaming
